view lock changes
