+++ views/view_lock.handlebars
@@ -1,27 +1,21 @@
 <div class="row">
   <div class="col-auto">
-    <h2>Welcome, {{name}}!</h2>
+    <h2>Hi, {{name}}!</h2>
+    <h3>Welcome to your Lockbox!</h3>
+
   </div>
 </div>
 
 <div class="row mt-4">
   <div class="col-md-6">
-    <h3>Create a New Project:</h3>
+    <h3>View locks</h3>
 
     <form class="form new-project-form">
       <div class="form-group">
-        <label for="project-name">project name:</label>
+        <label for="Choose from the list below"></label>
         <input class="form-input" type="text" id="project-name" name="project-name" />
       </div>
-      <div class="form-group">
-        <label for="project-funding">needed funding ($):</label>
-        <input class="form-input" type="number" id="project-funding" name="project-funding" />
-      </div>
-      <div class="form-group">
-        <label for="project-desc">description:</label>
-        <textarea class="form-input" id="project-desc" name="project-desc"></textarea>
-      </div>
-      <div class="form-group">
+      
         <button type="submit" class="btn btn-primary">Create</button>
       </div>
     </form>
@@ -45,4 +39,4 @@
   {{/if}}
 </div>
 
-<script src="./js/viewLock.js"></script>
+<script src="./js/viewLocks.js"></script>
